Refactor: change namespaces names in Tests/Basic
Due to the renaming of Tests/Utils to Tests/Basic, namespaces names are
changes to stay consistent.
